fix: service tests naming
