+++ src/test/java/com/example/roommate/tests/services/RoomDomainServiceTest.java
@@ -17,6 +17,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
+import java.util.List;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.*;
@@ -34,7 +35,7 @@ class RoomDomainServiceTest {
     RoomApplicationData differentRoom = new RoomApplicationData(differentRoomID, new RoomNumber("102"));
 
     @Test
-    @DisplayName("testing the addRoom function")
+    @DisplayName("executing addRoom with 1 room makes finAll() correctly containing that 1 room")
     void test_1() {
 
 
@@ -44,7 +45,7 @@ class RoomDomainServiceTest {
     }
 
     @Test
-    @DisplayName("testing the removeRoom function")
+    @DisplayName("adding 1 room followed by removing that 1 room makes findAll not contain that 1 room")
     void test_2_1() {
 
         roomDomainService.addRoom(room);
@@ -57,7 +58,7 @@ class RoomDomainServiceTest {
     }
 
     @Test
-    @DisplayName("testing the removeRoom function when the room is not in the RoomRepository")
+    @DisplayName("executing addRoom with 1 room followed by removing a different room with makes finAll() still correctly contain that 1 room")
     void test_2_2() {
 
 
@@ -69,7 +70,7 @@ class RoomDomainServiceTest {
     }
 
     @Test
-    @DisplayName("testing the getRooms function")
+    @DisplayName("getRooms properly contains both rooms added to it via add(room)")
     void test_3() {
         Room room1 = new Room(roomID, new RoomNumber("103"));
         Room room2 = new Room(differentRoomID, new RoomNumber("104"));
@@ -85,7 +86,7 @@ class RoomDomainServiceTest {
     }
 
     @Test
-    @DisplayName("testing the findRoomByID function")
+    @DisplayName("findRoomByID does not throw any exception if the room is already added AND also returns that room correctly")
     void test_4_1() throws NotFoundRepositoryException {
         roomDomainService.addRoom(room);
 
@@ -96,7 +97,7 @@ class RoomDomainServiceTest {
     }
 
     @Test
-    @DisplayName("testing the findRoomByID function when there is no room there")
+    @DisplayName("after adding room A, findRoomById of a different throws NotFoundRepositoryException as its not there")
     void test_4_2() {
 
         roomDomainService.addRoom(room);
@@ -104,29 +105,110 @@ class RoomDomainServiceTest {
         assertThatThrownBy(() -> roomDomainService.findRoomByID(differentRoomID)).isInstanceOf(NotFoundRepositoryException.class);
     }
 
+
+    @Test
+    @DisplayName("executing addRoom with room A makes findAll() contain room A")
+    void test_5() { //Problem: RoomEnrty.CalenderDay's are Objects, not comparable values like Strings or UUDI's
+
+
+        Room room = new Room(roomID, new RoomNumber("101"));
+        roomDomainService.addRoom(new RoomApplicationData(roomID, room.getRoomNumber()));
+
+        RoomEntry roomEntry = new RoomEntry(room.getRoomID(), room.getRoomNumber(), List.of());
+        assertThat(roomRepository.findAll().stream().map(IRoom::getRoomID))
+                .contains(roomEntry.roomID());
+    }
+
+    
+    
+
+   /* @Test //fixing tests by adding customAssertions
+    @DisplayName("testing the removeRoom function")
+    void test_2_1() {
+
+        Room room = new Room(id, "102");
+        roomService.addRoom(room);
+
+        roomService.removeRoom(room);
+
+
+        RoomEntry roomEntry = new RoomEntry(room.getRoomID(), room.getRoomNumber());
+        assertThat(roomRepository.findAll()).doesNotContain(roomEntry);
+    }
+
+    @Test
+    @DisplayName("testing the removeRoom function when the room is not in the RoomRepository")
+    void test_2_2() {
+        
+
+        Room room = new Room(id, "102");
+        Room differentRoom = new Room(differentRoomID, "102");
+
+        roomService.addRoom(room);
+
+        roomService.removeRoom(differentRoom);
+
+        RoomEntry roomEntry = new RoomEntry(room.getRoomID(), room.getRoomNumber());
+        assertThat(roomRepository.findAll()).contains(roomEntry);
+    }
+
+    @Test
+    @DisplayName("testing the getRooms function")
+    void test_3() {
+        Room room1 = new Room(id, "103");
+        Room room2 = new Room(differentRoomID, "104");
+        ArrayList<Room> rooms = new ArrayList<>();
+        rooms.add(room1);
+        rooms.add(room2);
+        RoomRepository roomRepository = new RoomRepository(rooms.stream().map(r -> new RoomEntry(r.getRoomID(), r.getRoomNumber())).toList());
+        RoomDomainService roomService = ServiceFactory.createRoomService(roomRepository, itemRepository);
+
+        //one room already exists
+        assertThat(roomService.getRooms().size()).isEqualTo(2);
+        assertThat(roomService.getRooms()).contains(room1, room2);
+    }
+
+    @Test
+    @DisplayName("testing the findRoomByID function")
+    void test_4_1() throws NotFoundRepositoryException {
+        Room room = new Room(id, "105");
+        roomService.addRoom(room);
+
+        assertThatCode(()->roomService.findRoomByID(room.getRoomID()))
+                .doesNotThrowAnyException();
+        assertThat(roomService.findRoomByID(room.getRoomID())).isEqualTo(room);
+    }
+
+    @Test
+    @DisplayName("testing the findRoomByID function when there is no room there")
+    void test_4_2() {
+
+        Room room = new Room(id, "105");
+        roomService.addRoom(room);
+
+        assertThatThrownBy(()->roomService.findRoomByID(differentRoomID)).isInstanceOf(NotFoundRepositoryException.class);
+    }
+
     @Test
     @DisplayName("testing the saveAll function")
-    @Disabled
     void test_5() {
-        Room room1 = new Room(roomID, new RoomNumber("106"));
-        Room room2 = new Room(differentRoomID, new RoomNumber("107"));
+        Room room1 = new Room(id, "106");
+        Room room2 = new Room(differentRoomID, "107");
 
-        //   roomDomainService.saveAll(List.of(room1, room2));
+        roomService.saveAll(List.of(room1, room2));
 
-        assertThat(roomRepository.findAll().stream().map(IRoom::getRoomID)).contains(room1.getRoomID(), room2.getRoomID());
+        assertThat(roomRepository.findAll().stream().map(IRoom::getRoomID)).contains(new RoomEntry(room1.getRoomID(), room1.getRoomNumber()), new RoomEntry(room2.getRoomID(), room2.getRoomNumber()));
     }
 
     @DisplayName("Adding a room that is already in the List does not change it")
     @Test
-    @Disabled
     void test_6() {
-        Room room = new Room(roomID, new RoomNumber("106"));
+        Room room = new Room(id, "106");
 
-        //  roomDomainService.saveAll(List.of(room));
-        roomDomainService.addRoom(new RoomApplicationData(roomID, room.getRoomNumber()));
+        roomService.saveAll(List.of(room));
+        roomService.addRoom(room);
 
-        assertThat(roomRepository.findAll().stream().map(IRoom::getRoomID)).containsOnlyOnce(room.getRoomID());
-    }
-
+        assertThat(roomRepository.findAll()).containsOnlyOnce(new RoomEntry(room.getRoomID(), room.getRoomNumber()));
+    }*/
 }
